Sensors, limit switches, etc.
Messed with sensors, then removed the code for them as they will not work.

Messed with limit switches, got them working, now they trigger intake in command to stop if it is running when button is pressed, and started work on getting that to start conveyer up, but did not have enough time to get scheduling done, so we have a sequentialcommandgroup that we used to get some examples, just ignore it.